forge fmt

+++ src/interfaces/IWETH.sol
@@ -6,5 +6,5 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 interface IWETH is IERC20 {
     function deposit() external payable;
 
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 }
